+++ CHANGELOG.md
@@ -18,6 +18,7 @@ New features / improvements:
 - Added new function `verb-body-lf-to-crlf` designed for use with requests sending multipart data.
 - Added new `Verb-Prelude` heading property, which can be used to specify an Emacs Lisp or JSON file to load variables from, before performing requests.
 - Added `verb-shell` and `verb-unix-epoch` utility functions.
+- Allow using Org [hyperlinks](https://orgmode.org/guide/Hyperlinks.html) in URLs, for example: `get [[http://example.com][my example]]`.
 
 ## **2.16.0** - 2024-03-02 (MELPA Stable)
 - Fixed LF being used instead of CRLF in multipart boundaries.